DemBas_num_decimales_px_IECA

Redondea un vector numérico para corregir el número de decimales que viene de un fichero px del IECA

Usage

DemBas_num_decimales_px_IECA(v1, num_decimales = 2)

Arguments

v1

vector numérico a redondear

num_decimales

número de decimales especificados en el fichero px de metadatos del IECA

Examples

df_19824 = DemBas_import_px_IECA("19824", "ieca_19824.px")
df_19824_meta = DemBas_leer_metadatos_px_IECA("ieca_19824.px")
df_19824_meta[df_19824_meta$Claves %in% c("DECIMALS", "SHOWDECIMALS"),]
#>          Claves Valores
#> 16     DECIMALS       2
#> 17 SHOWDECIMALS       2
df_19824$value = DemBas_num_decimales_px_IECA(df_19824$value)
head(df_19824,15)
#>                                       Medida Territorio     value
#> 1                      Grado de urbanización       Abla        NA
#> 2          % de población en centros urbanos       Abla   0.00000
#> 3     % de población en agrupaciones urbanas       Abla   0.00000
#> 4  % de población en celdas de malla rurales       Abla 100.00000
#> 5                      Grado de urbanización   Abrucena        NA
#> 6          % de población en centros urbanos   Abrucena   0.00000
#> 7     % de población en agrupaciones urbanas   Abrucena   0.00000
#> 8  % de población en celdas de malla rurales   Abrucena 100.00000
#> 9                      Grado de urbanización       Adra        NA
#> 10         % de población en centros urbanos       Adra   0.00000
#> 11    % de población en agrupaciones urbanas       Adra  80.49042
#> 12 % de población en celdas de malla rurales       Adra  19.50958
#> 13                     Grado de urbanización  Albanchez        NA
#> 14         % de población en centros urbanos  Albanchez   0.00000
#> 15    % de población en agrupaciones urbanas  Albanchez   0.00000