Calcula la tabla de vida para edades agrupadas pero simulando cálculos con una calculadora

Calcula la tabla de vida abreviada a partir de las tasas de mortalidad para edades agrupadas: 0,1,2,... y simula los cálculos con una calculadora

Usage

DemBas_tablavida_abreviada_calculadora(nMx, l0 = 1e+05, muestraSx = TRUE)

Arguments

l0

población inicial o radix de la tabla

muestraSx

si TRUE muestra la columna Sx

mx

vector de tasas de mortalidad para edades simples: 0,1,2,...

Value

devuelve un data.frame con la tabla de vida

Examples

(mx0 = 1733/441881) # TMI = D^t_0/N^t
#> [1] 0.00392187
# Defunciones de menores de un año durante 2003: 1733
# Nacimientos en España en 2003: 441881
mx = c(mx0,0.00027, 0.00013, 0.00016, 0.00043, 0.00057, 0.00059, 0.00081,
     0.00115, 0.00174, 0.00258, 0.00376, 0.00569, 0.00818, 0.01346,
     0.02206, 0.03844, 0.06981, 0.12872, 0.21674, 0.31705, 0.48258)
tv = DemBas_tablavida_abreviada_calculadora(mx)