Calcula las tasas especificas de mortalidad para edades agrupadas

Calcula las tasas especificas de mortalidad para edades agrupadas a partir de las defunciones y población

Usage

DemBas_mx_abreviada(Px, Dx, N0, D0)

Arguments

Px

vector de población media por edad

Dx

vector de defunciones por edad

N0

nacimientos en el periodo

D0

defunciones en el periodo

Value

devuelve un vector con las tasas especificas de mortalidad

Examples

x = c(0,1,seq(5,100,by=5))
Px = c(NA,1627456,1938350,2104636,2388049,3070467,3614444,3545550,
       3431304,3182840,2791972,2498361,2334676,1953022,1978465,
       1898370,1492487,974162,495260,203924,46078,5139)
Dx = c(NA,442,254,341,1032,1746,2136,2872,
       2933,5545,7193,9401,13294,15972,26636,
       41879,57377,68007,63007,44198,14609,2480)
N0 = 441881 # Nacimientos en España en 2003: 441881
D0 = 1733 # Defunciones de menores de un año durante 2003: 1733
mx = DemBas_mx(Px,Dx,N0,D0)
mx
#>       0       1       2       3       4       5       6       7       8       9 
#> 0.00392 0.00027 0.00013 0.00016 0.00043 0.00057 0.00059 0.00081 0.00085 0.00174 
#>      10      11      12      13      14      15      16      17      18      19 
#> 0.00258 0.00376 0.00569 0.00818 0.01346 0.02206 0.03844 0.06981 0.12722 0.21674 
#>      20     21+ 
#> 0.31705 0.48258